gf_rust

Module tile_collections

Implements synchronized loading of *.png representation of tiles into grid_forge on the Rust side and [TileSet] on the Godot side.

Parallel loading is made to make sure that the tile_type_id of [BasicIdentTileData] matches between [VisCollection] and [GodotTileMapCollection].

Structs

SingleTile

SingleTile is a resource that holds information about a single tile in a tilemap. It is used to load and save tiles from/to Godot’s TileMap.

TileCollections

TileCollections holds reference to chosen GodotTileset and and png tileset, gathering their information synchronously to create analogous [VisCollection] and [GodotTileMapCollection]