Struct gf_rust::tile_gen::TileGenerator

pub struct TileGenerator { /* private fields */ }

Class handling the generation of the tilemap.

The class is responsible for the generation of the [GridMap2D], and trasferring its results to the Godot’s [TileMap] afterwards.

The generation itself is done in a separate thread, so the main thread will not be blocked.

Implementations

impl TileGenerator

pub fn get_collection( &self ) -> <Option<Gd<TileCollections>> as GodotConvert>::Via

pub fn set_collection( &mut self, collection: <Option<Gd<TileCollections>> as GodotConvert>::Via )

pub fn get_modal(&self) -> <Option<Gd<AcceptDialog>> as GodotConvert>::Via

pub fn set_modal( &mut self, modal: <Option<Gd<AcceptDialog>> as GodotConvert>::Via )

pub fn get_pregenerated(&self) -> <Dictionary as GodotConvert>::Via

pub fn set_pregenerated( &mut self, pregenerated: <Dictionary as GodotConvert>::Via )

pub fn get_generation_time_us_total(&self) -> <u32 as GodotConvert>::Via

pub fn set_generation_time_us_total( &mut self, generation_time_us_total: <u32 as GodotConvert>::Via )

pub fn get_generation_time_us_success(&self) -> <u32 as GodotConvert>::Via

pub fn set_generation_time_us_success( &mut self, generation_time_us_success: <u32 as GodotConvert>::Via )